major cleanup of the major defensives file, added diffuse
