+++ src/analysis/retail/monk/brewmaster/modules/core/MajorDefensives/core.tsx
@@ -1,30 +1,15 @@
+import styled from '@emotion/styled';
+import { formatDuration, formatNumber } from 'common/format';
 import Spell from 'common/SPELLS/Spell';
-import talents from 'common/TALENTS/monk';
 import { Talent } from 'common/TALENTS/types';
-import { Section, SubSection, useInfo } from 'interface/guide';
-import Explanation from 'interface/guide/components/Explanation';
+import MAGIC_SCHOOLS, { color } from 'game/MAGIC_SCHOOLS';
+import { SpellLink, Tooltip } from 'interface';
 import Analyzer, { Options, SELECTED_PLAYER } from 'parser/core/Analyzer';
-import Events, {
-  AddStaggerEvent,
-  AnyEvent,
-  ApplyBuffEvent,
-  DamageEvent,
-  EventType,
-  RemoveBuffEvent,
-  RemoveStaggerEvent,
-} from 'parser/core/Events';
-import { GapHighlight } from 'parser/ui/CooldownBar';
-import CastEfficiencyBar from 'parser/ui/CastEfficiencyBar';
-import { ReactNode } from 'react';
+import Events, { AnyEvent, ApplyBuffEvent, DamageEvent, RemoveBuffEvent } from 'parser/core/Events';
+import BoringValue from 'parser/ui/BoringValueText';
 import Statistic from 'parser/ui/Statistic';
 import STATISTIC_CATEGORY from 'parser/ui/STATISTIC_CATEGORY';
-import { formatDuration, formatNumber } from 'common/format';
-import BoringValue from 'parser/ui/BoringValueText';
-import { SpellLink, Tooltip } from 'interface';
-import EventFilter from 'parser/core/EventFilter';
-import SPELLS from 'common/SPELLS';
-import MAGIC_SCHOOLS, { color } from 'game/MAGIC_SCHOOLS';
-import styled from '@emotion/styled';
+import { ReactNode } from 'react';
 
 export type MitigatedEvent = {
   event: AnyEvent;
@@ -217,151 +202,4 @@ export class MajorDefensive extends Analyzer {
       </Statistic>
     );
   }
-}
-
-const MAJOR_DEFENSIVES = [
-  talents.CELESTIAL_BREW_TALENT,
-  talents.FORTIFYING_BREW_TALENT,
-  talents.DAMPEN_HARM_TALENT,
-  talents.DIFFUSE_MAGIC_TALENT,
-  talents.ZEN_MEDITATION_TALENT,
-];
-
-export class ZenMeditation extends MajorDefensive {
-  constructor(options: Options) {
-    super({ talent: talents.ZEN_MEDITATION_TALENT }, options);
-
-    this.addEventListener(Events.damage.to(SELECTED_PLAYER), this.recordDamage);
-  }
-
-  private recordDamage(event: DamageEvent) {
-    if (this.defensiveActive && !event.sourceIsFriendly) {
-      this.recordMitigation({
-        event,
-        mitigatedAmount: absoluteMitigation(event, 0.6),
-      });
-    }
-  }
-}
-
-export class FortifyingBrew extends MajorDefensive {
-  private fortBrewStaggerPool: number = 0;
-  private hasGaiPlins = false;
-
-  constructor(options: Options) {
-    super(
-      { talent: talents.FORTIFYING_BREW_TALENT, buffSpell: SPELLS.FORTIFYING_BREW_BRM_BUFF },
-      options,
-    );
-
-    this.addEventListener(Events.damage.to(SELECTED_PLAYER), this.recordDamage);
-
-    if (this.selectedCombatant.hasTalent(talents.FORTIFYING_BREW_DETERMINATION_TALENT)) {
-      this.addEventListener(new EventFilter(EventType.AddStagger), this.recordStagger);
-      this.addEventListener(new EventFilter(EventType.RemoveStagger), this.recordPurify);
-
-      this.hasGaiPlins = this.selectedCombatant.hasTalent(talents.GAI_PLINS_IMPERIAL_BREW_TALENT);
-    }
-  }
-
-  private recordDamage(event: DamageEvent) {
-    if (this.defensiveActive && !event.sourceIsFriendly) {
-      this.recordMitigation({
-        event,
-        mitigatedAmount: absoluteMitigation(event, 0.2),
-      });
-    }
-  }
-
-  private recordStagger(event: AddStaggerEvent) {
-    if (this.defensiveActive) {
-      this.fortBrewStaggerPool += 0.15 * event.amount;
-    }
-  }
-
-  private recordPurify(event: RemoveStaggerEvent) {
-    // we don't use a fixed 50% for this because it could be mitigated by Staggering Strikes or one of the other related talents.
-    const purifyRatio = event.amount / (event.amount + event.newPooledDamage);
-    const purifyAmount = Math.ceil(purifyRatio * this.fortBrewStaggerPool);
-
-    if (this.defensiveActive && event.trigger?.ability.guid !== SPELLS.STAGGER_TAKEN.id) {
-      this.recordMitigation({
-        event,
-        mitigatedAmount: purifyAmount * (this.hasGaiPlins ? 1.25 : 1),
-      });
-    }
-
-    this.fortBrewStaggerPool = Math.max(0, this.fortBrewStaggerPool - purifyAmount);
-  }
-
-  mitigationSegments(mit: Mitigation): MitigationSegment[] {
-    const damage = mit.mitigated
-      .filter((event) => event.event.type === EventType.Damage)
-      .map((event) => event.mitigatedAmount)
-      .reduce((a, b) => a + b, 0);
-
-    const purifyBase = mit.mitigated
-      .filter((event) => event.event.type === EventType.RemoveStagger)
-      .map((event) => event.mitigatedAmount)
-      .reduce((a, b) => a + b, 0);
-
-    let purify = purifyBase;
-    let gaiPlins = 0;
-    if (this.hasGaiPlins) {
-      purify = purifyBase / 1.25;
-      gaiPlins = purifyBase - purify;
-    }
-
-    return [
-      {
-        amount: damage,
-        color: color(MAGIC_SCHOOLS.ids.PHYSICAL),
-        tooltip: (
-          <>
-            Base <SpellLink id={talents.FORTIFYING_BREW_TALENT} />
-          </>
-        ),
-      },
-      {
-        amount: purify,
-        color: 'rgb(112, 181, 112)',
-        tooltip: <SpellLink id={talents.FORTIFYING_BREW_DETERMINATION_TALENT} />,
-      },
-      {
-        amount: gaiPlins,
-        color: color(MAGIC_SCHOOLS.ids.HOLY),
-        tooltip: <SpellLink id={talents.GAI_PLINS_IMPERIAL_BREW_TALENT} />,
-      },
-    ].filter((seg) => seg.amount > 0);
-  }
-}
-
-export default function MajorDefensivesSection(): JSX.Element | null {
-  const info = useInfo();
-
-  if (!info) {
-    return null;
-  }
-
-  return (
-    <Section title="Major Defensives">
-      <Explanation>
-        In Dragonflight, Brewmaster Monk has gained multiple major defensive cooldowns. Using these
-        effectively is critical for your survival, especially while undergeared.
-      </Explanation>
-      <SubSection>
-        {MAJOR_DEFENSIVES.map(
-          (talent) =>
-            info.combatant.hasTalent(talent) && (
-              <CastEfficiencyBar
-                spellId={talent.id}
-                gapHighlightMode={GapHighlight.FullCooldown}
-                useThresholds
-                key={talent.id}
-              />
-            ),
-        )}
-      </SubSection>
-    </Section>
-  );
 }
